festival of function prototypes

+++ main_d2/collide.cpp
@@ -1150,7 +1150,7 @@ void apply_damage_to_controlcen(object* controlcen, fix damage, short who)
 		if (Game_mode & GM_MULTI) {
 			if (who == Players[Player_num].objnum)
 				add_points_to_score(CONTROL_CEN_SCORE);
-			multi_send_destroy_controlcen((ushort)(controlcen - Objects), Objects[who].id);
+			multi_send_destroy_controlcen((controlcen - Objects), Objects[who].id);
 		}
 #endif
